GHA: inline the GHA utilities (#23)
Inline the implementation for download and install.  This is motivated by the changes to the GHA infrastructure which changed the behaviour of the existing implementation. The event identifies the fragility due to the use of a non-stable interface. We now provide our own implementation. While this is complex, it does avoid the problem of unexpected behavioural changes. Even if we wanted to adopt the behaviour for the changes, we would not be able to provide a backwards compatible implementation which would make this flakey until we are firmly in the new implementation world.

Co-authored-by: Jeff <jeffdav@users.noreply.github.com>
Co-authored-by: Tristan Labelle <tristan@thebrowser.company>